Additional refactoring to support mobile launchpad
